[INJICERT-1277] Refactor code to initialise oauth related bean only if authorization module is certify
