+++ certify-service/src/main/java/io/mosip/certify/services/IarPresentationService.java
@@ -17,6 +17,7 @@ import io.mosip.certify.repository.IarSessionRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
@@ -40,6 +41,7 @@ import java.util.*;
  */
 @Slf4j
 @Service
+@ConditionalOnProperty(name = "mosip.certify.authorization-module", havingValue = "certify")
 public class IarPresentationService {
 
     @Autowired
