+++ certify-service/src/main/resources/application-local.properties
@@ -264,52 +264,55 @@ spring.jpa.hibernate.ddl-auto=none
 spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation=true
 spring.datasource.schema=./db_scripts/mosip_certify/ddl/certify-ca_cert_store.sql,./db_scripts/mosip_certify/ddl/certify-credential_template.sql,./db_scripts/mosip_certify/ddl/certify-key_alias.sql,./db_scripts/mosip_certify/ddl/certify-key_policy_def.sql,./db_scripts/mosip_certify/ddl/certify-key_store.sql,./db_scripts/mosip_certify/ddl/certify-rendering_template.sql,./db_scripts/mosip_certify/ddl/certify-credential_config.sql,./db_scripts/mosip_certify/dml/
 
+## ------------------------------------------- Authorization Server Configuration -------------------------------------------
+## set authorization-module as certify to use certify as authorization server
+mosip.certify.authorization-module=esignet
 ## ------------------------------------------- OAuth AS Metadata Configuration -------------------------------------------
 # Use below properties to use certify as authorization server
 #mosip.certify.authorization.url=http://localhost:8090
 #mosip.certify.authn.issuer-uri=${mosip.certify.domain.url}
 #mosip.certify.authn.jwk-set-uri=${mosip.certify.domain.url}${server.servlet.path}/oauth/.well-known/jwks.json
 #mosip.certify.authn.allowed-audiences={ '${mosip.certify.domain.url}/v1/certify/issuance/credential' }
-mosip.certify.oauth.issuer=${mosip.certify.domain.url}
-mosip.certify.oauth.token-endpoint=${mosip.certify.domain.url}/v1/certify/oauth/token
-mosip.certify.oauth.jwks-uri=${mosip.certify.domain.url}/v1/certify/oauth/.well-known/jwks.json
-mosip.certify.oauth.grant-types-supported=authorization_code
-mosip.certify.oauth.response-types-supported=code
-mosip.certify.oauth.code-challenge-methods-supported=S256
-mosip.certify.oauth.interactive-authorization-endpoint=${mosip.certify.domain.url}/v1/certify/oauth/iar
+#mosip.certify.oauth.issuer=${mosip.certify.domain.url}
+#mosip.certify.oauth.token-endpoint=${mosip.certify.domain.url}/v1/certify/oauth/token
+#mosip.certify.oauth.jwks-uri=${mosip.certify.domain.url}/v1/certify/oauth/.well-known/jwks.json
+#mosip.certify.oauth.grant-types-supported=authorization_code
+#mosip.certify.oauth.response-types-supported=code
+#mosip.certify.oauth.code-challenge-methods-supported=S256
+#mosip.certify.oauth.interactive-authorization-endpoint=${mosip.certify.domain.url}/v1/certify/oauth/iar
 
 ## ------------------------------------------- IAR (Interactive Authorization Request) Configuration -------------------------------------------
-mosip.certify.iar.identity-data=uin,vid,UIN,UID
+#mosip.certify.iar.identity-data=uin,vid,UIN,UID
 
 
-# OAuth Token Configuration
-mosip.certify.oauth.token.expires-in-seconds=3600
-mosip.certify.oauth.c-nonce.expires-in-seconds=300
-mosip.certify.oauth.token.type=Bearer
+## OAuth Token Configuration
+#mosip.certify.oauth.token.expires-in-seconds=3600
+#mosip.certify.oauth.c-nonce.expires-in-seconds=300
+#mosip.certify.oauth.token.type=Bearer
 
-# JWT Access Token Configuration
-mosip.certify.oauth.access-token.audience=${mosip.certify.identifier}/v1/certify/issuance/credential
+## JWT Access Token Configuration
+#mosip.certify.oauth.access-token.audience=${mosip.certify.identifier}/v1/certify/issuance/credential
 
-# Authorization Code Configuration
-mosip.certify.iar.authorization-code.expires-minutes=10
-mosip.certify.iar.authorization-code.length=24
+## Authorization Code Configuration
+#mosip.certify.iar.authorization-code.expires-minutes=10
+#mosip.certify.iar.authorization-code.length=24
 
-# Verify service integration
+## Verify service integration
 # Only configure the base URL and VP request endpoint - other endpoints are obtained dynamically from verify service
-mosip.certify.verify.service.base-url=http://localhost:8080
-mosip.certify.verify.service.vp-request-endpoint=${mosip.certify.verify.service.base-url}/v1/verify/vp-request
-mosip.certify.verify.service.vp-result-endpoint=${mosip.certify.verify.service.base-url}/v1/verify/vp-result
-mosip.certify.verify.service.verifier-client-id=certify-verifier-client
+#mosip.certify.verify.service.base-url=http://localhost:8080
+#mosip.certify.verify.service.vp-request-endpoint=${mosip.certify.verify.service.base-url}/v1/verify/vp-request
+#mosip.certify.verify.service.vp-result-endpoint=${mosip.certify.verify.service.base-url}/v1/verify/vp-result
+#mosip.certify.verify.service.verifier-client-id=certify-verifier-client
 
-# Presentation Definition
-mosip.certify.verify.service.presentation-definition.id=MOSIPVerifiableCredential
-mosip.certify.verify.service.presentation-definition.name=MOSIP ID
-mosip.certify.verify.service.presentation-definition.purpose=Relying party is requesting your digital ID for the purpose of Self-Authentication
+## Presentation Definition
+#mosip.certify.verify.service.presentation-definition.id=MOSIPVerifiableCredential
+#mosip.certify.verify.service.presentation-definition.name=MOSIP ID
+#mosip.certify.verify.service.presentation-definition.purpose=Relying party is requesting your digital ID for the purpose of Self-Authentication
 
-# Input Descriptor
-mosip.certify.verify.service.presentation-definition.input-descriptors[0].id=id card credential
+## Input Descriptor
+#mosip.certify.verify.service.presentation-definition.input-descriptors[0].id=id card credential
 
-# Constraints ? Fields
-mosip.certify.verify.service.presentation-definition.input-descriptors[0].constraints.fields[0].path[0]=$.type
-mosip.certify.verify.service.presentation-definition.input-descriptors[0].constraints.fields[0].filter.type=object
-mosip.certify.verify.service.presentation-definition.input-descriptors[0].constraints.fields[0].filter.pattern=MOSIPVerifiableCredential
+## Constraints ? Fields
+#mosip.certify.verify.service.presentation-definition.input-descriptors[0].constraints.fields[0].path[0]=$.type
+#mosip.certify.verify.service.presentation-definition.input-descriptors[0].constraints.fields[0].filter.type=object
+#mosip.certify.verify.service.presentation-definition.input-descriptors[0].constraints.fields[0].filter.pattern=MOSIPVerifiableCredential
